fix(monorepo): rename test files
