+++ packages/asset-registry/src/test/AssetConsumingFacade.test.ts
