+++ packages/asset-registry/src/test/AssetConsumingLogic.test.ts
