+++ packages/asset-registry/src/test/AssetContractLookup.test.ts
