+++ packages/asset-registry/src/test/AssetProducingFacade.test.ts
