+++ packages/asset-registry/src/test/AssetProducingLogic.test.ts
