+++ packages/erc-test-contracts/package.json
@@ -17,7 +17,7 @@
     "build": "yarn compile && yarn build-ts",
     "compile": "truffle compile",
     "deploy-contracts": "truffle migrate",
-    "test": "npm run build && mocha dist/js/src/test/ --timeout 90000",
+    "test": "npm run build && mocha dist/js/src/test/*.test.js --timeout 90000",
     "build-ts": "rm -rf dist/js && tsc -b tsconfig.build.json --verbose --pretty",
     "start-ganache": "ganache-cli -m 'chalk park staff buzz chair purchase wise oak receive avoid avoid home' -l 8000000 -e 1000000 -a 20",
     "prettier": "prettier --write --config-precedence file-override './src/**/*'",
