+++ packages/event-listener/package.json
@@ -20,7 +20,7 @@
         "lint-fix": "eslint 'src/**/*{.ts,.tsx}' --fix",
         "prettier": "prettier --write --config-precedence file-override './src/**/*'",
         "prepare": "npm run build",
-        "test": "rm -rf db.json && npm run build && mocha dist/js/test/ --timeout 60000 --exit",
+        "test": "rm -rf db.json && npm run build && mocha dist/js/test/*.test.js --timeout 60000 --exit",
         "test:concurrent": "concurrently --success first --kill-others -n eth,backend,test \"yarn start-ganache\" \"yarn start-test-backend\" \"yarn test\"",
         "start-test-backend": "node ../../node_modules/@energyweb/utils-testbackend/dist/js/src/index.js",
         "start-ganache": "ganache-cli -m 'chalk park staff buzz chair purchase wise oak receive avoid avoid home' -l 8000000 -e 1000000 -a 20",
