+++ packages/market/package.json
@@ -36,8 +36,8 @@
     "lint-fix": "solium -d contracts --fix",
     "start-ganache": "ganache-cli -m 'chalk park staff buzz chair purchase wise oak receive avoid avoid home' -l 8000000 -e 1000000 -a 20",
     "start-test-backend": "node ../../node_modules/@energyweb/utils-testbackend/dist/js/src/index.js",
-    "test": "mocha dist/js/src/test/ --timeout 60000 --exit",
-    "test-ts": "mocha -r ts-node/register src/test/*.ts --timeout 60000",
+    "test": "mocha dist/js/src/test/*.test.js --timeout 60000 --exit",
+    "test-ts": "mocha -r ts-node/register src/test/*.test.js --timeout 60000",
     "prettier": "prettier --write --config-precedence file-override './src/**/*'",
     "test-contracts": "concurrently --success first --kill-others -n eth,backend,test \"yarn start-ganache\" \"yarn start-test-backend\" \"wait-on tcp:8545 && yarn test\"",
     "clean": "rm -rf build dist schemas",
