+++ packages/user-registry/src/test/UserLogic.test.ts
