+++ packages/utils-general/package.json
@@ -30,7 +30,7 @@
     "prepare": "scripts/build.js",
     "start-ganache": "ganache-cli -m 'chalk park staff buzz chair purchase wise oak receive avoid avoid home' -g 0 -l 80000000 -e 1000000 -a 20",
     "prettier": "prettier --write --config-precedence file-override './src/**/*'",
-    "test-ts": "mocha -r ts-node/register test/*",
+    "test-ts": "mocha -r ts-node/register src/test/*.test.js",
     "clean": "rm -rf build dist",
     "build:static": "yarn compile"
   },
